requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
openai==1.51.0
numpy==1.24.3
faiss-cpu==1.7.4
rank-bm25==0.2.2
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
streamlit==1.28.1
python-multipart==0.0.6
tiktoken==0.5.1
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
